Cseplonorbert/bind comment to user

+++ server.py
@@ -116,26 +116,34 @@ def delete_answer(answer_id, question_id):
 
 @app.route("/question/<question_id>/new-comment", methods=["GET", "POST"])
 def add_question_comment(question_id):
-    if request.method == "POST":
-        question_comment = dict()
-        question_comment["message"] = request.form['comment']
-        question_comment["edited_count"] = 0
-        question_comment["question_id"] = question_id
-        data_manager.add_new_question_comment(question_comment)
+    if 'email' in session:
+        if request.method == "POST":
+            question_comment = dict()
+            question_comment["message"] = request.form['comment']
+            question_comment["edited_count"] = 0
+            question_comment["question_id"] = question_id
+            question_comment["user_id"] = session['user_id']
+            data_manager.add_new_question_comment(question_comment)
+            return redirect(url_for('display_question', question_id=question_id))
+        return render_template("add_question_comment.html", question_id=question_id)
+    else:
         return redirect(url_for('display_question', question_id=question_id))
-    return render_template("add_question_comment.html", question_id=question_id)
 
 
 @app.route("/answer/<answer_id>/<question_id>/new-comment", methods=["GET", "POST"])
 def add_answer_comment(answer_id, question_id):
-    if request.method == "POST":
-        answer_comment = dict()
-        answer_comment["message"] = request.form["comment"]
-        answer_comment['edited_count'] = 0
-        answer_comment['answer_id'] = answer_id
-        data_manager.add_answer_comment(answer_comment)
-        return redirect(f"/question/{question_id}")
-    return render_template("add_answer_comment.html", answer_id=answer_id, question_id=question_id)
+    if 'email' in session:
+        if request.method == "POST":
+            answer_comment = dict()
+            answer_comment["message"] = request.form["comment"]
+            answer_comment['edited_count'] = 0
+            answer_comment['answer_id'] = answer_id
+            answer_comment['user_id'] = session['user_id']
+            data_manager.add_answer_comment(answer_comment)
+            return redirect(url_for('display_question', question_id=question_id))
+        return render_template("add_answer_comment.html", answer_id=answer_id, question_id=question_id)
+    else:
+        return redirect(url_for('display_question', question_id=question_id))
 
 
 @app.route("/question/<question_id>/vote_up", methods=["GET"])
@@ -164,12 +172,15 @@ def answer_vote_down(answer_id, question_id):
 
 @app.route("/comment/<comment_id>/<question_id>/edit", methods=["GET", "POST"])
 def edit_comment(comment_id, question_id):
-    comment = data_manager.get_comment_by_id(comment_id)[0]
-    if request.method == "POST":
-        comment["message"] = request.form.get("comment")
-        data_manager.edit_comment(comment)
-        return redirect(url_for("display_question", question_id=question_id))
-    return render_template("edit_comment.html", comment=comment, question_id=question_id)
+    comment = data_manager.get_comment_by_id(comment_id)
+    if 'email' in session:
+        if session['user_id'] == comment['user_id']:
+            if request.method == "POST":
+                comment["message"] = request.form.get("comment")
+                data_manager.edit_comment(comment)
+            else:
+                return render_template("edit_comment.html", comment=comment, question_id=question_id)
+    return redirect(url_for("display_question", question_id=question_id))
 
 
 @app.route("/answer/<answer_id>/<question_id>/edit", methods=["GET", "POST"])
@@ -191,7 +202,10 @@ def edit_answer(answer_id, question_id):
 
 @app.route("/comments/<comment_id>/<question_id>delete")
 def delete_comment(comment_id, question_id):
-    data_manager.delete_comment(comment_id)
+    if 'email' in session:
+        comment = data_manager.get_comment_by_id(comment_id)
+        if comment['user_id'] == session['user_id']:
+            data_manager.delete_comment(comment_id)
     return redirect(url_for('display_question', question_id=question_id))
